add test to market
